Setup checker and remove redundant
