+++ package.json
@@ -11,22 +11,15 @@
   },
   "dependencies": {
     "@tma.js/sdk-react": "^2.2.2",
-    "@ton/phaser-sdk": "^0.0.1-beta.0",
-    "@twa-dev/sdk": "^7.0.0",
-    "@types/node": "^20.14.0",
-    "@vitejs/plugin-basic-ssl": "^1.1.0",
-    "clsx": "^2.1.1",
     "phaser": "^3.80.1",
-    "postcss": "^8.4.38",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-router-dom": "^6.23.1",
-    "react-scramble": "^0.5.1",
-    "tailwind-merge": "^2.3.0",
     "use-scramble": "^2.2.15",
     "zustand": "^4.5.2"
   },
   "devDependencies": {
+    "@types/node": "^20.14.0",
     "@types/react": "^18.2.66",
     "@types/react-dom": "^18.2.22",
     "@typescript-eslint/eslint-plugin": "^7.2.0",
@@ -36,8 +29,14 @@
     "eslint": "^8.57.0",
     "eslint-plugin-react-hooks": "^4.6.0",
     "eslint-plugin-react-refresh": "^0.4.6",
+    "prettier": "^3.3.0",
     "tailwindcss": "^3.4.3",
     "typescript": "^5.2.2",
-    "vite": "^5.2.0"
+    "vite": "^5.2.0",
+    "vite-plugin-checker": "^0.6.4",
+    "postcss": "^8.4.38",
+    "clsx": "^2.1.1",
+    "tailwind-merge": "^2.3.0",
+    "@vitejs/plugin-basic-ssl": "^1.1.0"
   }
 }
