+++ src/components/error-boundary-error.tsx
@@ -8,8 +8,8 @@ export const ErrorBoundaryError: FC<{ error: unknown }> = ({ error }) => (
         {error instanceof Error
           ? error.message
           : typeof error === "string"
-          ? error
-          : JSON.stringify(error)}
+            ? error
+            : JSON.stringify(error)}
       </code>
     </blockquote>
   </div>
