--- src/constants.ts
@@ -1,1 +1,0 @@
-export const OBJECTS_AMOUNT = 100;
